se finaliza el refactoring de clean arquitecture

+++ backend/BMT-backend/Domain/Entities/ProductDetails.cs
@@ -5,7 +5,7 @@
         public string ProductId { get; set; } = null!;
         public string ProductName { get; set; } = null!;
         public int Quantity { get; set; }
-        public int ProductsCost { get; set; }
+        public double ProductsCost { get; set; }
         public string EnterpriseName { get; set; } = null!;
         public string EnterpriseEmail { get; set; } = null!;
     }
